+++ app/src/main/java/com/example/trail/data/local/UserBookStateEntity.kt
@@ -6,7 +6,7 @@ import com.example.trail.data.ReadStatus
 
 @Entity(tableName = "userBookStateEntity")
 data class UserBookStateEntity(
-    @PrimaryKey val bookId: Int,
+    @PrimaryKey val bookId: Long,
     var readStatus: String = ReadStatus.NONE.name,
     var isFavourite: Boolean = false
 )
